third_party/blink/web_tests/plugins/sequential-focus.html

<!DOCTYPE html>
<body>
<script src="../resources/js-test.js"></script>

<input id="input-pre">
<iframe srcdoc="<object type='application/x-webkit-test-webplugin'></object>"></iframe>
<input id="input-post">

<script>
description('Test if sequential focus navigation works well with OBJECT in IFRAME.');
jsTestIsAsync = true;
window.onload = function() {
  setTimeout(function() {
    if (!window.eventSender) {
        debug('Manual test instruction: Focus on the first INPUT, press TAB twice, confirm the second INPUT gets focus.');
        return;
    }
    document.getElementById('input-pre').focus();
    eventSender.keyDown('\t');
    eventSender.keyDown('\t');
    shouldBeEqualToString('document.activeElement.id', 'input-post');
    finishJSTest();
  },50);
};
</script>
</body>
