third_party/blink/web_tests/plugins/sequential-focus-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
Test if sequential focus navigation works well with OBJECT in IFRAME.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement.id is "input-post"
PASS successfullyParsed is true

TEST COMPLETE

  
